Refactor and add more clasp open tests
